+++ iotdb-core/datanode/src/main/java/org/apache/iotdb/db/pipe/connector/thrift/IoTDBThriftConnector.java
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-package org.apache.iotdb.db.pipe.connector.base;
+package org.apache.iotdb.db.pipe.connector.thrift;
 
 import org.apache.iotdb.common.rpc.thrift.TEndPoint;
 import org.apache.iotdb.pipe.api.PipeConnector;
