+++ iotdb-core/datanode/src/main/java/org/apache/iotdb/db/pipe/connector/thrift/IoTDBThriftConnectorRequestVersion.java
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-package org.apache.iotdb.db.pipe.connector;
+package org.apache.iotdb.db.pipe.connector.thrift;
 
 public enum IoTDBThriftConnectorRequestVersion {
   VERSION_1((byte) 1),
