+++ iotdb-core/datanode/src/main/java/org/apache/iotdb/db/pipe/connector/thrift/payload/reponse/PipeTransferFilePieceResp.java
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-package org.apache.iotdb.db.pipe.connector.v1.reponse;
+package org.apache.iotdb.db.pipe.connector.thrift.payload.reponse;
 
 import org.apache.iotdb.common.rpc.thrift.TSStatus;
 import org.apache.iotdb.service.rpc.thrift.TPipeTransferResp;
