+++ iotdb-core/datanode/src/main/java/org/apache/iotdb/db/pipe/connector/thrift/payload/request/PipeTransferTabletReq.java
@@ -17,12 +17,12 @@
  * under the License.
  */
 
-package org.apache.iotdb.db.pipe.connector.v1.request;
+package org.apache.iotdb.db.pipe.connector.thrift.payload.request;
 
 import org.apache.iotdb.commons.exception.MetadataException;
 import org.apache.iotdb.commons.utils.PathUtils;
-import org.apache.iotdb.db.pipe.connector.IoTDBThriftConnectorRequestVersion;
-import org.apache.iotdb.db.pipe.connector.v1.PipeRequestType;
+import org.apache.iotdb.db.pipe.connector.thrift.IoTDBThriftConnectorRequestVersion;
+import org.apache.iotdb.db.pipe.connector.thrift.payload.PipeRequestType;
 import org.apache.iotdb.db.queryengine.plan.parser.StatementGenerator;
 import org.apache.iotdb.db.queryengine.plan.statement.crud.InsertTabletStatement;
 import org.apache.iotdb.service.rpc.thrift.TPipeTransferReq;
