+++ iotdb-core/datanode/src/main/java/org/apache/iotdb/db/pipe/connector/thrift/sync/IoTDBThriftConnectorClient.java
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-package org.apache.iotdb.db.pipe.connector.v1;
+package org.apache.iotdb.db.pipe.connector.thrift.sync;
 
 import org.apache.iotdb.commons.client.ThriftClient;
 import org.apache.iotdb.commons.client.property.ThriftClientProperty;
